try GCN again

+++ model/temporal/temporal_ext_gcn.py
@@ -95,11 +95,11 @@ class TemporalExtGCN(nn.Module):
     def forward(self, x):
         node_x, edge_idx, edge_attr = x.x, x.edge_index, x.edge_attr
 
-        node_x = node_x.double()
+        node_x = node_x.float()
         edge_idx = edge_idx
         edge_attr = edge_attr
 
-        print("temp_ext_gcn node_x:", node_x.shape, type(node_x.double()), node_x.double().dtype)
+        print("temp_ext_gcn node_x:", node_x.shape, type(node_x), node_x.dtype)
         print("temp_ext_gcn edge_idx:", edge_idx.shape, type(edge_idx), edge_idx.dtype)
         print("temp_ext_gcn edge_attr:", edge_attr.shape, type(edge_attr), edge_attr.dtype)
